Hack until template looks like the old site
This was fun. Rapid iteration times are amazing for motivation.

Adjust the font sizes to be larger. Maybe not to everyone's taste, but
I'm finding myself sitting further back from the screen as of late, and
reading them is more comfortable. Maybe I'm just older? Maybe I'll
change my mind next time and find it rediculously large. I think the
reason that it's comfortable is that it's clamping the line length. I
wonder why there's a sweet spot for line length? Probably don't want it
too long so that the eye can move vertically for context; something
along the lines of limited memory.

Not sure if Bootstrap 4 has moved on since I cobbled this together (it's
still using Bootstrap 3).